Added more items
I've added more items for the player to interact with and spice levels up with (specifically the launcher, NME, coins, and a chest to break). Tweaked more of the existing elements, like the code for the player's speed changing their color and how that's registered by other items they need to break by moving fast enough.